pyproject.toml
[tool.poetry]
name = "labels"
version = "0.1.0"
description = ""
authors = ["Viktor Stískala <viktor@stiskala.cz>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
PyYAML = "^6.0"

[tool.poetry.scripts]
generate-labels = 'labels.generator:generate_labels'

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
